Make display updates more efficient by only updating changed pages.
Instead of writing the entire buffer every time the display contents
change, we can update just the pages that had been changed. This means
that (in tests, using a slow USB I2C bridge) the updates of the
display are now much faster if only parts of the display change,
as the transaction overhead was very high. Previously a full update
of the diplay was around 0.71s on my test system with a 128x64
display. For most updates which only affect a single page, this is
now reduced to 0.095s, increasing as the number of pages changed
increases. With the shortcut for unchanged buffer, the display
function can now take next to no time.